+++ common/include/pcl/range_image/impl/range_image_planar.hpp
@@ -94,7 +94,7 @@ RangeImagePlanar::calculate3DPoint (double image_x, double image_y, double range
   //cout << __PRETTY_FUNCTION__ << " called.\n";
   double delta_x = (image_x+static_cast<double> (image_offset_x_)-center_x_)*focal_length_x_reciprocal_,
         delta_y = (image_y+static_cast<double> (image_offset_y_)-center_y_)*focal_length_y_reciprocal_;
-  point[2] = range / (sqrtf (delta_x*delta_x + delta_y*delta_y + 1));
+  point[2] = range / (sqrt (delta_x*delta_x + delta_y*delta_y + 1));
   point[0] = delta_x*point[2];
   point[1] = delta_y*point[2];
   point = to_world_system_ * point;
@@ -107,7 +107,7 @@ RangeImagePlanar::getImagePoint (const Eigen::Vector3d& point, double& image_x, 
   Eigen::Vector3d transformedPoint = to_range_image_system_ * point;
   if (transformedPoint[2]<=0)  // Behind the observer?
   {
-    image_x = image_y = range = -1.0f;
+    image_x = image_y = range = -1.0;
     return;
   }
   range = transformedPoint.norm ();
