+++ common/src/gaussian.cpp
@@ -45,12 +45,12 @@ pcl::GaussianKernel::compute (double sigma,
   assert (kernel_width %2 == 1);
   assert (sigma >= 0);
   kernel.resize (kernel_width);
-  static const double factor = 0.01f;
-  static const double max_gauss = 1.0f;
+  static const double factor = 0.01;
+  static const double max_gauss = 1.0;
   const int hw = kernel_width / 2;
-  double sigma_sqr = 1.0f / (2.0f * sigma * sigma);
+  double sigma_sqr = 1.0 / (2.0 * sigma * sigma);
   for (int i = -hw, j = 0, k = kernel_width - 1; i < 0 ; i++, j++, k--)
-    kernel[k] = kernel[j] = expf (-static_cast<double>(i) * static_cast<double>(i) * sigma_sqr);
+    kernel[k] = kernel[j] = exp (-static_cast<double>(i) * static_cast<double>(i) * sigma_sqr);
   kernel[hw] = 1;
   unsigned g_width = kernel_width;
   for (unsigned i = 0; fabs (kernel[i]/max_gauss) < factor; i++, g_width-= 2) ;
@@ -82,14 +82,14 @@ pcl::GaussianKernel::compute (double sigma,
   assert (sigma >= 0);
   kernel.resize (kernel_width);
   derivative.resize (kernel_width);
-  const double factor = 0.01f;
-  double max_gauss = 1.0f, max_deriv = double (sigma * exp (-0.5));
+  const double factor = 0.01;
+  double max_gauss = 1.0, max_deriv = double (sigma * exp (-0.5));
   int hw = kernel_width / 2;
 
-  double sigma_sqr = 1.0f / (2.0f * sigma * sigma);
+  double sigma_sqr = 1.0 / (2.0 * sigma * sigma);
   for (int i = -hw, j = 0, k = kernel_width - 1; i < 0 ; i++, j++, k--)
   {
-    kernel[k] = kernel[j] = expf (-static_cast<double>(i) * static_cast<double>(i) * sigma_sqr);
+    kernel[k] = kernel[j] = exp (-static_cast<double>(i) * static_cast<double>(i) * sigma_sqr);
     derivative[j] = -static_cast<double>(i) * kernel[j];
     derivative[k] = -derivative[j];
   }
